Rename hook files (#79)
