+++ cxflow/hooks/stop_after.py
